chore: cleanup pass — Decimal money, single PrismaClient, services layer
#1 Float → Decimal for monetary fields (Invoice.grossAmount/paidAmount,
   Product.pricePLN/priceEUR, Deal.value, ConsignmentItem.unitPrice).
   Float caused silent precision drift (e.g. paid==gross compares failing
   on rounded values). All read-side arithmetic and JSON responses now wrap
   the field with Number() so downstream callers keep getting numbers, not
   Decimal-stringified objects. Note: GET responses for Deal and Product
   may now serialize prices as JSON strings if not explicitly converted —
   only affects callers that do math on the raw response.

#2 Removed --accept-data-loss from npm start. Future schema changes that
   would silently drop columns or narrow types will now error out at deploy
   instead of destroying data without warning.

#3 Single shared PrismaClient via src/db.js. Previously index.js, inbox-poller
   and mail-sender each instantiated their own client (3 connection pools);
   now one pool, consistent with how routes already access prisma.

#4 Extracted services layer:
   - src/services/contractor-match.js  (scoreContractor, normalizeContractorName)
   - src/services/ifirma-sync.js       (processIfirmaInvoices, guessCountryFromInv)
   Removed routes/contractors.js → routes/invoices.js cross-import.
   routes/contractors.js now exports only `router`, consistent with siblings.

+++ src/routes/analytics.js
@@ -57,8 +57,8 @@ const SCHEMA_DESCRIPTION = `SCHEMA BAZY DANYCH SurfStickBell:
 Table: Invoice
 - id (UUID), number (String, np "25/2026"), contractorId (UUID FK),
 - ifirmaId (Int), issueDate (DateTime), dueDate (DateTime),
-- grossAmount (Float), currency (String: "EUR"/"PLN"),
-- paidAmount (Float), status (String: "paid"/"unpaid"/"partial"),
+- grossAmount (Decimal(12,2)), currency (String: "EUR"/"PLN"),
+- paidAmount (Decimal(12,2)), status (String: "paid"/"unpaid"/"partial"),
 - type (String), ifirmaType (String, np "prz_dostawa_ue_towarow" = WDT),
 - source (String: "system"/"ifirma_sync"), bankReference (String),
 - extras (JSONB: matchedOrder, kontrahentNazwa, ...)
